requirements.txt
asgiref==3.4.1
blis==0.2.4
certifi==2021.5.30
charset-normalizer==2.0.1
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
click==8.0.1
cymem==2.0.5
Django==3.2.5
django-cors-headers==3.7.0
django-extensions==3.1.3
djangorestframework==3.12.4
idna==3.2
joblib==1.0.1
mathparse==0.1.2
murmurhash==1.0.5
mysqlclient==2.0.3
nltk==3.6.2
numpy==1.21.0
packaging==21.0
Pint==0.17
plac==0.9.6
preshed==2.0.1
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.7.5
python-dotenv==0.18.0
pytz==2021.1
PyYAML==3.13
regex==2021.7.6
requests==2.26.0
six==1.16.0
spacy==2.1.9
SQLAlchemy==1.2.19
sqlparse==0.4.1
srsly==1.0.5
thinc==7.0.8
tqdm==4.61.2
urllib3==1.26.6
wasabi==0.8.2
